TelephonyProvider: use name as part of unique apn identifier
This allows carrier APN updates to propogate to the currently selected
APN. Before it would create a new APN with the same name and the old one
would be selected, leaving the user in a bad state.

Now since names are unique it would update as necessary, preserving user
selection with the new values.

Ticket: PAELLA-294

Change-Id: Iec3da72043b7596536d7de16e59a1cd99e81d9fe
Signed-off-by: Roman Birg <roman@cyngn.com>